BaseEntity added

+++ src/main/java/com/avitam/fantasy11/model/UserRepository.java
@@ -2,5 +2,5 @@ package com.avitam.fantasy11.model;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface UserRepository extends MongoRepository<User, Long> {
+public interface UserRepository extends MongoRepository<User, Integer> {
 }
